requirements.txt
streamlit
langchain
langchain-openai
tavily-python
pandas
python-dotenv
pydantic